+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Redirect user into `/` after checkin login failure (@mkasztelnik)
 - Service categories (@mkasztelnik)
 - Service order possibility (@mkasztelnik)
+- Configure `database_cleaner` for rspec (@mkasztelnik)
 
 ### Changed